backend/requirements.txt
sqlalchemy
PyMySQL>=0.8.1,<0.9
Pillow
boto3
jinja2
